+++ CHANGELOG.md
@@ -4,6 +4,11 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## [Unreleased]
+### Added
+- Added `renderSortOrder` to `Sprite3D`.
+- Added `lerp` to `Point3D`.
+
 ## [2.1.1] - 2022-12-23
 ### Fixed
 - Fixed an issue which caused some meshes not to render when using vertex colors.